Spawn windows in selected client's cwd
The existing spawn_cwd patch requires setting up your shell to print
its current working directory to the terminal title, and it parses that
title using strtok(3), strchr(3), and malloc(3). That seems like an ugly
hack to me.

The proc(5) file system symlinks each process' working directories
to /proc/<pid>/cwd. Simply calling chdir(2) with this symlink before
calling execvp(2) should produce the same effect in theory. However,
this will not work for terminals (probably the most common use case),
because the shell that the terminal spawned changes its working directory,
not the terminal itself.

This patch provides a function, cdtochildwd, which locates the children of
the given process, and then chdirs into (currently) the working directory
of the first child it finds. This allows dwm to find the working directory
of bash that was spawned by st for example.

In the future, recursion or looping could be used to find the working
directory of, for example, the deepest process in the tree. This would
allow the patch to work with arbitrarily deep stack of processes (e.g.,
st→bash→nix develop→lf→etc.)